+++ client/java/src/main/java/io/openlineage/client/transports/Transport.java
@@ -13,6 +13,7 @@ import lombok.NonNull;
 public abstract class Transport {
   enum Type {
     CONSOLE,
+    FILE,
     HTTP,
     KAFKA,
     KINESIS,