+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   *Extract dataset schema from `KafkaSource` when `GenericRecord` deserialized is used.*
 * **Flink: support Flink version 1.17.1** [`#1947`](https://github.com/OpenLineage/OpenLineage/pull/1947) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)  
   *Support Flink versions: 1.15.4, 1.16.2 and 1.17.1.*
+* **Spark: Databricks platform integration test.** [`#1928`](https://github.com/OpenLineage/OpenLineage/pull/1928) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)  
+  *Spark integration test to verify behaviour on databricks platform to be run manually in CircleCI when needed.*
 
 ## [0.28.0](https://github.com/OpenLineage/OpenLineage/compare/0.27.2...0.28.0) - 2023-06-12
 ### Added